still got a _lot_ of stuff to work out, have to finish re-working the pwm library to take full 16bit data for timer 1, and still don't have a good way of seeing if data that should be getting read from the eeprom on my mcp4725 is working or not, besides that got i2c workng for talking to the DAC on the mcp4725 but again not sure if the eeprom is or not

+++ pwm.c
@@ -11,6 +11,8 @@ freenode/#linuxandsci - JoshAshby
 #include "adc.h"
 #include "pwm.h"
 #include "global.h"
+#include "i2c.h"
+
 void pwm_setup_all(void){
     TCCR0B |= (1<<CS00)
             | (1<<CS01);
@@ -41,12 +43,12 @@ void pwm_setup0(void)
     DDRD |= (1<<5);
     DDRD |= (1<<6);
 }
-void pwm0A(unsigned int value)//set the duty cycle on the PWM
+void pwm0A(uint16_t value)//set the duty cycle on the PWM
 {
     TCCR0A |= (1<<COM0A1);
     OCR0A = value;
 }
-void pwm0B(unsigned int value)//set the duty cycle on the PWM
+void pwm0B(uint16_t value)//set the duty cycle on the PWM
 {
     TCCR0A |= (1<<COM0B1);
     OCR0B = value;
@@ -60,12 +62,12 @@ void pwm_setup1(void)
     DDRB |= (1<<1);
     DDRB |= (1<<2);
 }
-void pwm1A(int value)//set the duty cycle on the PWM
+void pwm1A(uint16_t value)//set the duty cycle on the PWM
 {
     TCCR1A |= (1<<COM1A1);
     OCR1A = value;
 }
-void pwm1B(unsigned int value)//set the duty cycle on the PWM
+void pwm1B(uint16_t value)//set the duty cycle on the PWM
 {
     TCCR1A |= (1<<COM1B1);
     OCR1B = value;
@@ -78,12 +80,12 @@ void pwm_setup2(void)
     DDRD |= (1<<3);
     DDRB |= (1<<3);
 }
-void pwm2A(unsigned int value)//set the duty cycle on the PWM
+void pwm2A(uint16_t value)//set the duty cycle on the PWM
 {
     TCCR2A |= (1<<COM2A1);
     OCR2A = value;
 }
-void pwm2B(unsigned int value)//set the duty cycle on the PWM
+void pwm2B(uint16_t value)//set the duty cycle on the PWM
 {
     TCCR2A |= (1<<COM2B1);
     OCR2B = value;
